refactor: norm 맞추기#25
함수 5개 맞추기 위해 is_valid_map.c 생성
