+++ srcs/parse_map/Makefile
@@ -6,7 +6,7 @@
 #    By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/04/27 15:13:41 by kipark            #+#    #+#              #
-#    Updated: 2022/11/07 21:12:26 by jayoon           ###   ########.fr        #
+#    Updated: 2022/11/10 20:36:19 by jayoon           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -41,7 +41,8 @@ SRCS_PARSE_MAP				:=	init_map_info.c			\
 								get_cnt_map_width.c		\
 								pass_empty_line.c		\
 								pass_element.c			\
-								get_map.c
+								get_map.c				\
+								is_valid_map.c
 
 SRCS_ERROR					:=	error.c
