+++ srcs/parse_map/asset/map_in_pdf.cub
@@ -26,10 +26,17 @@ WE ./path_to_the_west_texture
 111111111011000001110000000000001
 100000000011000001110111111111111
 11110111111111011100000010001
-11110111111110111010100100001
+11110111111110111010100100001111
 11000000110101011100000010001
 10000000000000001100000010001
 10000000000000001101010010001
 110000011101010111110111100N1
 11110111 1110101 101111010011
 11111111 1111111 111111111111
+ 
+11111
+ 1001
+ 1001
+ 1001
+ 1001
+11111
